refactor: Renombrar carpetas
